[wip][fix] replace addr type u32 to u64

+++ src/bus.rs
@@ -31,7 +31,7 @@ impl Bus {
     }
 
     // get 1 byte
-    pub fn raw_byte(&self, addr: u32) -> u8 {
+    pub fn raw_byte(&self, addr: u64) -> u8 {
         if self.mrom.in_range(addr) {
             self.mrom.raw_byte(addr)
         } else if self.clint.in_range(addr) {
@@ -44,7 +44,7 @@ impl Bus {
     }
 
     // store
-    pub fn store8(&mut self, addr: u32, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    pub fn store8(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
             self.mrom.store8(addr, data)
         } else if self.clint.in_range(addr) {
@@ -62,7 +62,7 @@ impl Bus {
 
     pub fn store16(
         &mut self,
-        addr: u32,
+        addr: u64,
         data: u32,
     ) -> Result<(), (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
@@ -82,7 +82,7 @@ impl Bus {
 
     pub fn store32(
         &mut self,
-        addr: u32,
+        addr: u64,
         data: u32,
     ) -> Result<(), (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
@@ -102,7 +102,7 @@ impl Bus {
 
     pub fn store64(
         &mut self,
-        addr: u32,
+        addr: u64,
         data: i64,
     ) -> Result<(), (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
@@ -121,7 +121,7 @@ impl Bus {
     }
 
     // load
-    pub fn load8(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    pub fn load8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
             self.mrom.load8(addr)
         } else if self.clint.in_range(addr) {
@@ -137,7 +137,7 @@ impl Bus {
         }
     }
 
-    pub fn load16(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    pub fn load16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
             self.mrom.load16(addr)
         } else if self.clint.in_range(addr) {
@@ -153,7 +153,7 @@ impl Bus {
         }
     }
 
-    pub fn load32(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    pub fn load32(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
             self.mrom.load32(addr)
         } else if self.clint.in_range(addr) {
@@ -169,7 +169,7 @@ impl Bus {
         }
     }
 
-    pub fn load64(&self, addr: u32) -> Result<u64, (Option<u32>, TrapCause, String)> {
+    pub fn load64(&self, addr: u64) -> Result<u64, (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
             self.mrom.load64(addr)
         } else if self.clint.in_range(addr) {
@@ -185,7 +185,7 @@ impl Bus {
         }
     }
 
-    pub fn load_u8(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    pub fn load_u8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
             self.mrom.load_u8(addr)
         } else if self.clint.in_range(addr) {
@@ -201,7 +201,7 @@ impl Bus {
         }
     }
 
-    pub fn load_u16(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    pub fn load_u16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         if self.mrom.in_range(addr) {
             self.mrom.load_u16(addr)
         } else if self.clint.in_range(addr) {
@@ -219,17 +219,17 @@ impl Bus {
 }
 
 pub trait Device {
-    fn in_range(&self, addr: u32) -> bool;
-    fn addr2index(&self, addr: u32) -> usize;
-    fn raw_byte(&self, addr: u32) -> u8;
-    fn store8(&mut self, addr: u32, data: u32) -> Result<(), (Option<u32>, TrapCause, String)>;
-    fn store16(&mut self, addr: u32, data: u32) -> Result<(), (Option<u32>, TrapCause, String)>;
-    fn store32(&mut self, addr: u32, data: u32) -> Result<(), (Option<u32>, TrapCause, String)>;
-    fn store64(&mut self, addr: u32, data: i64) -> Result<(), (Option<u32>, TrapCause, String)>;
-    fn load8(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)>;
-    fn load16(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)>;
-    fn load32(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)>;
-    fn load64(&self, addr: u32) -> Result<u64, (Option<u32>, TrapCause, String)>;
-    fn load_u8(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)>;
-    fn load_u16(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)>;
+    fn in_range(&self, addr: u64) -> bool;
+    fn addr2index(&self, addr: u64) -> usize;
+    fn raw_byte(&self, addr: u64) -> u8;
+    fn store8(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)>;
+    fn store16(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)>;
+    fn store32(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)>;
+    fn store64(&mut self, addr: u64, data: i64) -> Result<(), (Option<u32>, TrapCause, String)>;
+    fn load8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)>;
+    fn load16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)>;
+    fn load32(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)>;
+    fn load64(&self, addr: u64) -> Result<u64, (Option<u32>, TrapCause, String)>;
+    fn load_u8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)>;
+    fn load_u16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)>;
 }
