+++ src/bus/mrom.rs
@@ -3,7 +3,7 @@ use crate::TrapCause;
 
 pub struct Mrom {
     pub mrom: Vec<u8>,
-    pub base_addr: u32,
+    pub base_addr: u64,
     size: usize,
 }
 
@@ -43,23 +43,23 @@ impl Mrom {
 #[allow(clippy::identity_op)]
 impl Device for Mrom {
     // is addr in device address space
-    fn in_range(&self, addr: u32) -> bool {
+    fn in_range(&self, addr: u64) -> bool {
         (self.base_addr..=self.base_addr + self.size as u32).contains(&addr)
     }
 
     // address to raw index
-    fn addr2index(&self, addr: u32) -> usize {
+    fn addr2index(&self, addr: u64) -> usize {
         (addr - self.base_addr) as usize
     }
 
     // get 1 byte
-    fn raw_byte(&self, addr: u32) -> u8 {
+    fn raw_byte(&self, addr: u64) -> u8 {
         let addr = self.addr2index(addr);
         self.mrom[addr]
     }
 
     // store
-    fn store8(&mut self, addr: u32, _data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store8(&mut self, addr: u64, _data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
         Err((
             Some(addr),
             TrapCause::StoreAMOPageFault,
@@ -67,7 +67,7 @@ impl Device for Mrom {
         ))
     }
 
-    fn store16(&mut self, addr: u32, _data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store16(&mut self, addr: u64, _data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
         Err((
             Some(addr),
             TrapCause::StoreAMOPageFault,
@@ -75,7 +75,7 @@ impl Device for Mrom {
         ))
     }
 
-    fn store32(&mut self, addr: u32, _data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store32(&mut self, addr: u64, _data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
         Err((
             Some(addr),
             TrapCause::StoreAMOPageFault,
@@ -83,7 +83,7 @@ impl Device for Mrom {
         ))
     }
 
-    fn store64(&mut self, addr: u32, _data: i64) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store64(&mut self, addr: u64, _data: i64) -> Result<(), (Option<u32>, TrapCause, String)> {
         Err((
             Some(addr),
             TrapCause::StoreAMOPageFault,
@@ -92,17 +92,17 @@ impl Device for Mrom {
     }
 
     // load
-    fn load8(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let addr = self.addr2index(addr);
         Ok(self.mrom[addr] as i8 as i32 as u32)
     }
 
-    fn load16(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let addr = self.addr2index(addr);
         Ok(((self.mrom[addr + 1] as i16) << 8 | (self.mrom[addr + 0] as i16)) as i32 as u32)
     }
 
-    fn load32(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load32(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let addr = self.addr2index(addr);
         Ok((self.mrom[addr + 3] as u32) << 24
             | (self.mrom[addr + 2] as u32) << 16
@@ -110,7 +110,7 @@ impl Device for Mrom {
             | (self.mrom[addr + 0] as u32))
     }
 
-    fn load64(&self, addr: u32) -> Result<u64, (Option<u32>, TrapCause, String)> {
+    fn load64(&self, addr: u64) -> Result<u64, (Option<u32>, TrapCause, String)> {
         let addr = self.addr2index(addr);
         Ok((self.mrom[addr + 7] as u64) << 56
             | (self.mrom[addr + 6] as u64) << 48
@@ -121,12 +121,12 @@ impl Device for Mrom {
             | (self.mrom[addr + 1] as u64) << 8
             | (self.mrom[addr + 0] as u64))
     }
-    fn load_u8(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load_u8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let addr = self.addr2index(addr);
         Ok(self.mrom[addr] as u32)
     }
 
-    fn load_u16(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load_u16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let addr = self.addr2index(addr);
         Ok((self.mrom[addr + 1] as u32) << 8 | (self.mrom[addr + 0] as u32))
     }
