+++ src/cpu/fetch.rs
@@ -15,10 +15,10 @@ pub fn fetch(cpu: &mut Cpu) -> Result<Box<dyn Decode>, (Option<u64>, TrapCause, 
     } else {
         let index_pc2: u64 = cpu.trans_addr(TransFor::Fetch, TransAlign::Size8, cpu.pc + 2)?;
         log::infoln!("pc: 0x{:08x}", cpu.pc);
-        let new_inst: u64 = (cpu.bus.raw_byte(index_pc2 + 1) as u64) << 24
-            | (cpu.bus.raw_byte(index_pc2) as u64) << 16
-            | (cpu.bus.raw_byte(index_pc + 1) as u64) << 8
-            | (cpu.bus.raw_byte(index_pc) as u64);
+        let new_inst: u32 = (cpu.bus.raw_byte(index_pc2 + 1) as u32) << 24
+            | (cpu.bus.raw_byte(index_pc2) as u32) << 16
+            | (cpu.bus.raw_byte(index_pc + 1) as u32) << 8
+            | (cpu.bus.raw_byte(index_pc) as u32);
         Ok(Box::new(new_inst))
     }
 }
